# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog Post API Documentation",
        "description": "API documentation for Blog Post application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://395c-2001-ee0-262-1fc-84cc-955-2049-79ea.ngrok-free.app",
            "description": "Ngrok Tunnel Server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get List Users",
                "operationId": "26b68341036642eed7c31c7ca9dafa03",
                "parameters": [
                    {
                        "name": "ngrok-skip-browser-warning",
                        "in": "header",
                        "description": "Required for ngrok requests",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "true"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User By Id",
                "operationId": "8242536d482e67ca1a4607f836f473fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "User"
        }
    ],
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "API key needed to access the endpoints",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "apiKey": []
        }
    ]
}